third_party/blink/web_tests/external/wpt/custom-elements/revamped-scoped-registry/Element-customElementRegistry-exceptions.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] customElementRegistry on a failed custom element created by calling createElement on CustomElementRegistry should return the registry
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] customElementRegistry on a failed custom element created by setting innerHTML should return the associated scoped registry
  assert_equals: expected object "[object CustomElementRegistry]" but got null
[FAIL] customElementRegistry on a failed custom element created by parser should return the specified custom regsitry
  assert_equals: expected object "[object CustomElementRegistry]" but got null
Harness: the test ran to completion.
